Initial commit
The inital set up and work for the project, added in the assets, created the terrain and lay out, textured it, imported all neccesary assets needed for map design.

Scripted the movement of both the enemy goblins and the players builder character, phase one of the project is basically finished, could use alittle more refinement.