Big UI Updated
Added in the UI and programmed it for the gold, wave tracker, enemies remaing trackers. added in the ui and programmed the start round button with multiple audios attached. added in the main menu scene, styled it, added music and added in functioning buttons to start the game or exit.