Fixed minion hp ui, added base dmg
fixed the ui on the minions health bar so now it works properly, added health to the base wall and gave it a ui section. added now so when minions hit the wall it reduces the HP, plays a SFX and screenshakes the camera, if base health is reduced to 0 the game ends and game over UI panel opens.